Revision: rhythmbox-devel@gnome.org--2004/rhythmbox--main--0.7--patch-33
automatically adapt ratings (Michael Terry)

+++ widgets/rb-cell-renderer-rating.h
@@ -48,7 +48,7 @@ typedef struct
 {
 	GtkCellRendererClass parent_class;
 
-	void (*rated) (GtkTreePath *path, int rating);
+	void (*rated) (GtkTreePath *path, double rating);
 } RBCellRendererRatingClass;
 
 GtkType          rb_cell_renderer_rating_get_type (void);
